Move renaming in dedicated files
